% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_ID.R
\name{SSID_4ptFilter}
\alias{SSID_4ptFilter}
\title{Four Point SSID Filter}
\usage{
SSID_4ptFilter(data, tcrit_u, tcrit_l, n, ewma, stp)
}
\arguments{
\item{data}{Vector containing PV to be analyzed}

\item{tcrit_u}{The upper limit for the t-statistic, probable TS when returned t exceeds}

\item{tcrit_l}{The lower limit for the t-statistic, probable SS when returned t is below}

\item{n}{The window size for analysis}

\item{ewma}{The exponential weighted average filter value (typ. 0.1)}

\item{stp}{Step size (typ. 1)}
}
\value{
Dataframe containing T-statistics and SS indicator value
}
\description{
identifies probable SS in a PV, for use in either process engineering or economic data
}
\examples{
SSID_4ptFilter <- (data = ProcessData$Variable, tcrit_u = 3.2, tcrit_l = 1.0, n = 10, ewma = 0.07, stp = 1);
}
